+++ app/common/CenteredContentLayout.tsx
@@ -7,11 +7,9 @@ import { Layout } from './Layout';
  * Once children do not fit, vertical scroll will appear normally.
  */
 export const CenteredContentLayout: FC<PropsWithChildren> = ({ children }) => (
-  <div className="flex items-center min-h-full">
-    <Layout>
-      <div className="mx-auto xl:w-3/5 lg:w-3/4">
-        {children}
-      </div>
-    </Layout>
-  </div>
+  <Layout className="flex items-center justify-center">
+    <div className="m-auto w-full xl:w-3/5 lg:w-3/4">
+      {children}
+    </div>
+  </Layout>
 );